pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Python library for HeliosCTA Power Fundies"
authors = ["Aidan Keaveny <aidan.keaveny@helioscta.com>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
psycopg2-binary = ">=2.9"
pyodbc = ">=5.1"

python-dotenv = ">=1.1"
numpy = ">=2.3"
pandas = ">=2.3"
scipy = ">=1.16"
plotly = ">=6.3"